exercise 4.22

+++ controllers/blogs.js
@@ -1,13 +1,10 @@
 const blogsRouter = require('express').Router()
-const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const userExtractor = require('../utils/middleware').userExtractor
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
-
         .find({}).populate('user', { username: 1, name: 1 })
-
     response.json(blogs)
 })
 
@@ -20,10 +17,9 @@ blogsRouter.get('/:id', async (request, response) => {
     }
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
     const body = request.body
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
 
     const blog = new Blog({
         title: body.title,
@@ -39,18 +35,17 @@ blogsRouter.post('/', async (request, response) => {
     response.json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    const user = await User.findById(decodedToken.id)
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+    const user = request.user
     const blog = await Blog.findById(request.params.id)
-
-    if (blog.user.toString() === user._id.toString()){
-        await Blog.findByIdAndRemove(request.params.id)
-        response.status(204).end()
-    } else {
+    if (blog){
+        if (blog.user.toString() === user._id.toString()){
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        }
+    }else {
         response.status(404).end()
     }
-
 })
 
 blogsRouter.put('/:id', async (request, response) => {
